cli: Fix rendering of long integers
Recent changes to the human-readable rendering of outputs and console
values led to long integer values being presented in scientific
notation (e.g. 1.2345678e+07). While technically correct, this is an
unusual way to present integer values.

This commit checks if a number value is an integer, and if so renders it
as a sequence of digits instead (e.g. 12345678).